requirements.txt
sense-hat
paho-mqtt
Flask
psutil
PyYAML
